karma_bot/requirements.txt
python-telegram-bot==20.7
pytz==2024.1
python-dotenv
sqlite3
